Excel/VBA Sample UI
Adding image of proprietary Excel/VBA UI using distinct, beautiful UI/UX experience. All data originates in this file using individual tabs for Customers, Payments, Vehicles and Work Orders, allowing user to export to TSV. This is only an image, the actual Excel file is not available in this project.